Исправлена отказоустойчивость для etl
