+++ services/ugc/src/requirements.txt
@@ -1,7 +1,6 @@
 fastapi==0.104.0
 pydantic==2.4.2
 pydantic_settings==2.0.3
-kafka-python==2.0.2
 
 uvicorn==0.12.2
 gunicorn
